requirements.txt
python-telegram-bot>=21.3
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
rapidfuzz>=3.0.0